% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.cell}
\alias{data.cell}
\title{data.cell}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 121 rows and 21 columns.
}
\usage{
data.cell
}
\description{
The `data.cell` represents the second type of feature matrix used for calculating correlations (in this case, cell abundances), where rows correspond to samples, and columns correspond to features.
}
\examples{
library(iPRISM)
data(data.cell, package = "iPRISM")
dim(data.cell)

}
\keyword{datasets}
