% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.path}
\alias{data.path}
\title{data.path}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 121 rows and 17 columns.
}
\usage{
data.path
}
\description{
The `data.path` represents the first type of feature matrix used for calculating correlations (in this case, pathway expression levels), where rows correspond to samples, and columns correspond to features.
}
\examples{
library(iPRISM)
data(data.path, package = "iPRISM")
plot(data.path)

}
\keyword{datasets}
