% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sig}
\alias{data_sig}
\title{data_sig}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 121 rows and 31 columns.
}
\usage{
data_sig
}
\description{
The `data_sig` represents the sample feature matrix, where rows correspond to samples, and columns correspond to features.
}
\examples{
library(iPRISM)
data(data_sig, package = "iPRISM")
plot(data_sig)

}
\keyword{datasets}
