% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genelist_cp}
\alias{genelist_cp}
\title{TME gene list after random walks}
\format{
An object of class \code{numeric} of length 15867.
}
\usage{
genelist_cp
}
\description{
This gene list includes genes from tumor microenvironment (TME). Random Walk with Restart (RWR) is applied to prioritize genes that are relevant to immunotherapy responses.
}
\examples{
library(iPRISM)
data(genelist_cp, package = "iPRISM")

}
\keyword{datasets}
