% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genelist_hla}
\alias{genelist_hla}
\title{HLA gene list after random walks}
\format{
An object of class \code{numeric} of length 15867.
}
\usage{
genelist_hla
}
\description{
This gene list includes genes from human leukocyte antigen (HLA). Random Walk with Restart (RWR) is applied to prioritize genes that are relevant to immunotherapy responses.
}
\examples{
library(iPRISM)
data(genelist_hla, package = "iPRISM")

}
\keyword{datasets}
