% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genelist_imm}
\alias{genelist_imm}
\title{ICI gene list after random walks}
\format{
An object of class \code{numeric} of length 15867.
}
\usage{
genelist_imm
}
\description{
This gene list includes genes from immune checkpoint inhibitors (ICI). Random Walk with Restart (RWR) is applied to prioritize genes that are relevant to immunotherapy responses.
}
\examples{
library(iPRISM)
data(genelist_imm, package = "iPRISM")

}
\keyword{datasets}
