% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPRISM-package.R
\name{iPRISM-package}
\alias{iPRISM-package}
\alias{iPRISM}
\title{Intelligent Predicting Response to Cancer Immunotherapy Through Systematic Modeling}
\description{
Immunotherapy has revolutionized cancer treatment, but predicting patient
response remains challenging. Here, we presented Intelligent Predicting
Response to cancer Immunotherapy through Systematic Modeling (iPRISM), a
novel network-based model that integrates multiple data types to predict
immunotherapy outcomes. It incorporates gene expression, biological
functional network, tumor microenvironment characteristics, immune-related
pathways, and clinical data to provide a comprehensive view of factors
influencing immunotherapy efficacy. By identifying key genetic and
immunological factors, it provides an insight for more personalized
treatment strategies and combination therapies to overcome resistance
mechanisms.
}
\details{
iPRISM
}
\examples{
# if (!"devtools" \%in\% as.data.frame(installed.packages())$Package)
#   install.packages("devtools")
# devtools::install_github("hanjunwei-lab/iPRISM")

}
