% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_list}
\alias{path_list}
\title{path_list}
\format{
An object of class \code{list} of length 2656.
}
\usage{
path_list
}
\description{
The `path_list` contains the gene list associated with pathways.
}
\examples{
library(iPRISM)
data(path_list, package = "iPRISM")
length(path_list)

}
\keyword{datasets}
