% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pred_value}
\alias{pred_value}
\title{Original Class Labels for Samples}
\format{
An object of class \code{character} of length 121.

An object of class \code{character} of length 121.
}
\usage{
pred_value

pred_value
}
\description{
A named vector where each element corresponds to a sample name and represents the original class label.

A named vector where each element corresponds to a sample name and represents the original class label.
}
\examples{
library(iPRISM)
data(pred_value, package = "iPRISM")
table(pred_value)

library(iPRISM)
data(pred_value, package = "iPRISM")
table(pred_value)

}
\keyword{datasets}
