% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seeds}
\alias{Seeds}
\title{Seed Node Names}
\format{
An object of class \code{character} of length 3.
}
\usage{
Seeds
}
\description{
A character vector with seed node names.
}
\examples{
library(iPRISM)
data(Seeds, package = "iPRISM")

}
\keyword{datasets}
